+++ docs/changelog.md
@@ -6,6 +6,9 @@ Requires libvips v8.7.0.
 
 #### v0.21.2 - TBD
 
+* Ensure shortest edge is at least one pixel after resizing.
+  [#1003](https://github.com/lovell/sharp/issues/1003)
+
 * Expose `pages` and `pageHeight` metadata for multi-page input images.
   [#1205](https://github.com/lovell/sharp/issues/1205)